% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.programs.R
\docType{data}
\name{available.programs}
\alias{available.programs}
\title{Programs with microdata available for download}
\format{An object of class \code{data.frame} with 12 rows and 7 columns.}
\usage{
available.programs
}
\description{
This is a list of INEP programs (censuses, national exams and nation wide surveys)
whose microdata were made available by INEP for download
}
\keyword{datasets}
