% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.archives.R
\name{fetch.archives}
\alias{fetch.archives}
\title{Fetches ZIP files for a specific program}
\usage{
fetch.archives(program, years)
}
\arguments{
\item{program}{any value in \code{available.programs$program}}

\item{years}{any set of values in \code{available.programs$years[[program]]}

[[program]: R:[program}
}
\description{
Fetches ZIP files for a specific program
}
\examples{
fetch.archives("enem", 2000)
fetch.archives("enade", 2005:2008)
}
