\name{inepdata-package}
\alias{inepdata-package}
\alias{inepdata}
\docType{package}
\title{
\packageTitle{inepdata}
}
\description{
\packageDescription{inepdata}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{inepdata}
\packageIndices{inepdata}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{inepdata}

Maintainer: \packageMaintainer{inepdata}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
