% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.programs.R
\name{load.programs}
\alias{load.programs}
\title{Loads available programs}
\usage{
load.programs()
}
\description{
Loads programs available in download page given by \code{download.page.url} and available
locally as zip files given by \code{zip.path}
}
