% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{Options}
\alias{Options}
\title{Options for package \code{inepdata}}
\usage{
Options(..., zip.path, download.page.url, temp.path, keep.download, verbose)
}
\arguments{
\item{...}{these dots are here just to force parameters to be explicitly named}

\item{zip.path}{path to directory where the downloaded zip files are located,
so that those zip files will not be downloaded if available locally.
\strong{Important}: the filenames must be equal to the corresponding file in INEP site.
If you are not going to work with already downloaded ZIP files,
then set \code{zip.path} to "".}

\item{download.page.url}{external HTML page from which the ZIP URLs will be scraped.
It is there just in case the package fails to fetch the correct URL
from INEP's site -- which can occur when INEP alters its site structure,
for instance.
(Should you detect this kind of occurrence, please contact this package's author,
just in case he haven't noticed that yet.)
If you do not want to seek any download page, then set \code{download.page.url} to "".
If you want \code{download.page.url} back to its default value,
then set \code{download.page.url} to \code{NULL}.}

\item{temp.path}{where should the microdata ZIP files be downloaded and decompressed?
It cannot be set "" and, if attempted, it is set to its default value "./temp".}

\item{keep.download}{whether to keep or purge the downloaded packed microdata files}

\item{verbose}{do you want know what is going on under the hood while these functions
are running? (currently, not implemented)}
}
\description{
Options for package \code{inepdata}
}
\examples{
# in case INEP microdata download page changes:
options(download.page.url = "http://inep.hypothetical.new.page/microdata")
# if you have part of your data already downloaded
options(zip.path = "./zipfiles.I.already.downloaded/")
# if you only want to work with the data you already downloaded, without looking at INPE's page
options(download.page.url = "", zip.path = "./zipfiles.I.already.downloaded/") #
options(temp.path = "/tmp/")     # Unix example
options(keep.download = TRUE, temp.path = "./store.zip.files.here/")

}
