% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enem.R
\name{enem}
\alias{enem}
\title{Gets Data from ENEM (Exame Nacional de Ensino Médio)}
\usage{
enem(years)
}
\arguments{
\item{years}{years to get from INEP site or from the directory informed in \code{zip.files.path} below;
it should be a vector of either integers or strings}
}
\value{
A list of lists, indexed by the years present in \code{year} that could be processed.

For each year in \code{years}, either one list is returned in case of success (total or
partial) or \code{NULL} is returned in case of complete lack of success.

In case of success, the item for each year will be a list containing the following
fields:
\itemize{
\item \code{last.successful.step} -- a string; either "download", "unpack", "import", "organize" or "format"
\item \code{download.url} -- URL used to download the packed archive
\item \code{data.frame} -- what you are really looking for; it will be here in case of full
success; it will be either one single data frame or a list of data frames or \code{NULL}
}

There might be one or more of the following additional fields in the following
conditions:
\itemize{
\item If \code{keep.download == TRUE} or if \code{last.successful.step == download},
\itemize{
\item \code{downloaded.file} -- path and name of the downloaded packed file
}
\item If \code{last.successful.step == unpack},
\itemize{
\item \code{unpacked.files} -- path of the directory where the archive was unpacked
}
\item If \code{last.successful.step == import},
\itemize{
\item \code{raw.data.frames} -- it will be either one single data frame or a list of data frames
}
}
}
\description{
Downloads, unpacks, imports, organizes and formats microdata from ENEM editions
}
\examples{
# getting data from years 2014 to 2016
enem(2014:2016)

}
